+++ src/transaction/transaction.php
@@ -104,6 +104,7 @@ class Transaction extends Resource
         - before [DateTime or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
         - tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
         - externalIds [list of strings, default null]: list of external ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Project object, default null]: Project object. Not necessary if StarkBank\User.setDefaut() was set before function call
 
     ## Return: