+++ src/transfer/transfer.php
@@ -147,6 +147,7 @@ class Transfer extends Resource
         - taxId [string, default null]: filter for transfers sent to the specified tax ID. ex: "012.345.678-90"
         - sort [string, default "-created"]: sort order considered in response. Valid options are 'created', '-created', 'updated' or '-updated'.
         - tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
+        - ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - user [Project object, default null]: Project object. Not necessary if StarkBank\User.setDefaut() was set before function call
 
     ## Return: