+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- ids parameter to transaction.query
+- ids parameter to transfer.query
+
+### Fixed
+- UtilityPayment test case
 
 ## [0.6.0] - 2020-08-20
 ### Added